Modify: 리스트와 리스트아이템을 하나로 합침
떡국 카드 컴포넌트의 이름을 TteokgukCardsWithCaption으로 변경
리스트와 리스트아이템을 분리하지 않고 하나의 컴포넌트로 변경
