--- src/components/MyActivityPage/ReceivedIngredientsList.tsx
@@ -1,19 +1,0 @@
-import ReceivedIngredientsListItem from "./ReceivedIngredientsListItem";
-
-const ReceivedIngredientsList = () => {
-  return (
-    <ul>
-      <ReceivedIngredientsListItem />
-      <ReceivedIngredientsListItem />
-      <ReceivedIngredientsListItem />
-      <ReceivedIngredientsListItem />
-      <ReceivedIngredientsListItem />
-      <ReceivedIngredientsListItem />
-      <ReceivedIngredientsListItem />
-      <ReceivedIngredientsListItem />
-      <ReceivedIngredientsListItem />
-    </ul>
-  );
-};
-
-export default ReceivedIngredientsList;
