--- src/components/common/Tteokguks/TteokgukList.tsx
@@ -1,23 +1,0 @@
-import { css } from "@styled-system/css";
-
-import TteokgukCard from "./TteokgukCard";
-
-const TteokgukList = () => {
-  return (
-    <ul className={styles.container}>
-      {[...Array(12)].map(() => (
-        <TteokgukCard tteokgukNumber={1178} nickname="재민" />
-      ))}
-    </ul>
-  );
-};
-
-export default TteokgukList;
-
-const styles = {
-  container: css({
-    display: "flex",
-    flexFlow: "row wrap",
-    gap: "1.6rem",
-  }),
-};
