+++ src/pages/MyPage.tsx
@@ -5,8 +5,8 @@ import { css } from "@styled-system/css";
 import { Link } from "@/routes/Link";
 import Header from "@/components/common/Header";
 import IconButton from "@/components/common/IconButton";
-import TteokgukList from "@/components/Mypage/TteokgukList";
 import IngredientList from "@/components/Mypage/IngredientList";
+import TteokgukList from "@/components/common/TteokgukList";
 import GuideIcon from "@/assets/svg/profile.svg";
 import DumplingIcon from "@/assets/svg/dumpling.svg";
 import VisitIcon from "@/assets/svg/visit.svg";
