+++ src/pages/RandomUserPage.tsx
@@ -4,11 +4,11 @@ import { css } from "@styled-system/css";
 
 import Header from "@/components/common/Header";
 import IconButton from "@/components/common/IconButton";
+import UserProfileSection from "@/components/common/UserProfileSection";
 import TteokgukList from "@/components/common/TteokgukList";
 import GuideIcon from "@/assets/svg/profile.svg";
 import DumplingIcon from "@/assets/svg/dumpling.svg";
 import VisitIcon from "@/assets/svg/visit.svg";
-import UserProfileSection from "@/components/common/UserProfileSection";
 
 const RandomUserPage = () => {
   return (
